+++ app/src/main/kotlin/com/simplemobiletools/gallery/extensions/activity.kt
@@ -25,7 +25,7 @@ import com.simplemobiletools.gallery.R
 import com.simplemobiletools.gallery.activities.SimpleActivity
 import com.simplemobiletools.gallery.helpers.NOMEDIA
 import com.simplemobiletools.gallery.helpers.REQUEST_EDIT_IMAGE
-import com.simplemobiletools.gallery.helpers.REQUEST_SET_WALLPAPER
+import com.simplemobiletools.gallery.helpers.REQUEST_SET_AS
 import com.simplemobiletools.gallery.models.Directory
 import com.simplemobiletools.gallery.models.Medium
 import com.simplemobiletools.gallery.views.MySquareImageView
@@ -71,32 +71,32 @@ fun Activity.shareMedia(media: List<Medium>) {
     }
 }
 
-fun Activity.trySetAsWallpaper(file: File) {
+fun Activity.trySetAs(file: File) {
     try {
         var uri = Uri.fromFile(file)
-        if (!setAsWallpaper(uri, file)) {
+        if (!setAs(uri, file)) {
             uri = getFileContentUri(file)
-            setAsWallpaper(uri, file, false)
+            setAs(uri, file, false)
         }
     } catch (e: Exception) {
         toast(R.string.unknown_error_occurred)
     }
 }
 
-fun Activity.setAsWallpaper(uri: Uri, file: File, showToast: Boolean = true): Boolean {
+fun Activity.setAs(uri: Uri, file: File, showToast: Boolean = true): Boolean {
     var success = false
     Intent().apply {
         action = Intent.ACTION_ATTACH_DATA
         setDataAndType(uri, file.getMimeType("image/*"))
         addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-        val chooser = Intent.createChooser(this, getString(R.string.set_as_wallpaper_with))
+        val chooser = Intent.createChooser(this, getString(R.string.set_as))
 
         if (resolveActivity(packageManager) != null) {
-            startActivityForResult(chooser, REQUEST_SET_WALLPAPER)
+            startActivityForResult(chooser, REQUEST_SET_AS)
             success = true
         } else {
             if (showToast) {
-                toast(R.string.no_wallpaper_setter_found)
+                toast(R.string.no_capable_app_found)
             }
             success = false
         }
@@ -130,9 +130,9 @@ fun Activity.openWith(file: File, forceChooser: Boolean = true) {
         action = Intent.ACTION_VIEW
         setDataAndType(uri, file.getMimeType())
         addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-        val chooser = Intent.createChooser(this, getString(R.string.open_with))
 
         if (resolveActivity(packageManager) != null) {
+            val chooser = Intent.createChooser(this, getString(R.string.open_with))
             startActivity(if (forceChooser) chooser else this)
         } else {
             toast(R.string.no_app_found)
@@ -140,15 +140,19 @@ fun Activity.openWith(file: File, forceChooser: Boolean = true) {
     }
 }
 
-fun Activity.openEditor(file: File) {
-    val uri = Uri.fromFile(file)
+fun Activity.openFileEditor(file: File) {
+    openEditor(Uri.fromFile(file))
+}
+
+fun Activity.openEditor(uri: Uri, forceChooser: Boolean = false) {
     Intent().apply {
         action = Intent.ACTION_EDIT
         setDataAndType(uri, "image/*")
         addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
 
         if (resolveActivity(packageManager) != null) {
-            startActivityForResult(this, REQUEST_EDIT_IMAGE)
+            val chooser = Intent.createChooser(this, getString(R.string.edit_image_with))
+            startActivityForResult(if (forceChooser) chooser else this, REQUEST_EDIT_IMAGE)
         } else {
             toast(R.string.no_editor_found)
         }
@@ -260,7 +264,8 @@ fun SimpleActivity.toggleFileVisibility(oldFile: File, hide: Boolean, callback: 
 
 fun Activity.getFileSignature(path: String) = StringSignature(File(path).lastModified().toString())
 
-fun Activity.loadImage(path: String, target: MySquareImageView) {
+fun Activity.loadImage(path: String, target: MySquareImageView, verticalScroll: Boolean) {
+    target.isVerticalScrolling = verticalScroll
     if (path.isImageFast() || path.isVideoFast()) {
         if (path.isPng()) {
             loadPng(path, target)
@@ -304,7 +309,7 @@ fun Activity.loadAnimatedGif(path: String, target: MySquareImageView) {
             .load(path)
             .asGif()
             .signature(getFileSignature(path))
-            .diskCacheStrategy(DiskCacheStrategy.SOURCE)
+            .diskCacheStrategy(DiskCacheStrategy.NONE)
             .crossFade()
 
     if (config.cropThumbnails) builder.centerCrop() else builder.fitCenter()