+++ app/src/main/kotlin/com/simplemobiletools/gallery/extensions/context.kt
@@ -2,6 +2,7 @@ package com.simplemobiletools.gallery.extensions
 
 import android.content.Context
 import android.content.Intent
+import android.content.res.Configuration
 import android.database.Cursor
 import android.net.Uri
 import android.provider.MediaStore
@@ -18,6 +19,8 @@ import com.simplemobiletools.gallery.helpers.VIDEOS
 import com.simplemobiletools.gallery.models.Medium
 import java.io.File
 import java.util.*
+
+val Context.portrait get() = resources.configuration.orientation == Configuration.ORIENTATION_PORTRAIT
 
 fun Context.getRealPathFromURI(uri: Uri): String? {
     var cursor: Cursor? = null
@@ -139,35 +142,33 @@ private fun parseCursor(context: Context, cur: Cursor, isPickImage: Boolean, isP
         }
     }
 
-    if (curPath.isEmpty()) {
-        config.includedFolders.mapNotNull { File(it).listFiles() }.forEach {
-            for (file in it) {
-                val size = file.length()
-                if (size <= 0L) {
-                    continue
-                }
-
-                val filename = file.name
-                val isImage = filename.isImageFast() || filename.isGif()
-                val isVideo = if (isImage) false else filename.isVideoFast()
-
-                if (!isImage && !isVideo)
-                    continue
-
-                if (isVideo && (isPickImage || showMedia == IMAGES))
-                    continue
-
-                if (isImage && (isPickVideo || showMedia == VIDEOS))
-                    continue
-
-                val dateTaken = file.lastModified()
-                val dateModified = file.lastModified()
-
-                val medium = Medium(filename, file.absolutePath, isVideo, dateModified, dateTaken, size)
-                val isAlreadyAdded = curMedia.any { it.path == file.absolutePath }
-                if (!isAlreadyAdded)
-                    curMedia.add(medium)
+    config.includedFolders.filter { it.isEmpty() || it == curPath }.mapNotNull { File(it).listFiles() }.forEach {
+        for (file in it) {
+            val size = file.length()
+            if (size <= 0L) {
+                continue
             }
+
+            val filename = file.name
+            val isImage = filename.isImageFast() || filename.isGif()
+            val isVideo = if (isImage) false else filename.isVideoFast()
+
+            if (!isImage && !isVideo)
+                continue
+
+            if (isVideo && (isPickImage || showMedia == IMAGES))
+                continue
+
+            if (isImage && (isPickVideo || showMedia == VIDEOS))
+                continue
+
+            val dateTaken = file.lastModified()
+            val dateModified = file.lastModified()
+
+            val medium = Medium(filename, file.absolutePath, isVideo, dateModified, dateTaken, size)
+            val isAlreadyAdded = curMedia.any { it.path == file.absolutePath }
+            if (!isAlreadyAdded)
+                curMedia.add(medium)
         }
     }