+++ app/src/main/kotlin/com/simplemobiletools/gallery/fragments/VideoFragment.kt
@@ -95,6 +95,9 @@ class VideoFragment : ViewPagerFragment(), SurfaceHolder.Callback, SeekBar.OnSee
         super.setMenuVisibility(menuVisible)
         mIsFragmentVisible = menuVisible
         if (menuVisible) {
+            if (mSurfaceView != null)
+                initMediaPlayer()
+
             if (context?.config?.autoplayVideos == true) {
                 playVideo()
             }
@@ -228,6 +231,7 @@ class VideoFragment : ViewPagerFragment(), SurfaceHolder.Callback, SeekBar.OnSee
             }
         } catch (e: IOException) {
             Log.e(TAG, "init media player failed $e")
+            releaseMediaPlayer()
         }
     }
 
@@ -258,10 +262,14 @@ class VideoFragment : ViewPagerFragment(), SurfaceHolder.Callback, SeekBar.OnSee
     private fun cleanup() {
         pauseVideo()
         mCurrTimeView?.text = 0.getFormattedDuration()
-        mMediaPlayer?.release()
-        mMediaPlayer = null
+        releaseMediaPlayer()
         mSeekBar?.progress = 0
         mTimerHandler?.removeCallbacksAndMessages(null)
+    }
+
+    private fun releaseMediaPlayer() {
+        mMediaPlayer?.release()
+        mMediaPlayer = null
     }
 
     private fun videoPrepared(mediaPlayer: MediaPlayer) {
@@ -294,8 +302,7 @@ class VideoFragment : ViewPagerFragment(), SurfaceHolder.Callback, SeekBar.OnSee
     }
 
     override fun surfaceDestroyed(holder: SurfaceHolder) {
-        mMediaPlayer?.release()
-        mMediaPlayer = null
+        releaseMediaPlayer()
     }
 
     private fun setVideoSize() {
@@ -348,6 +355,9 @@ class VideoFragment : ViewPagerFragment(), SurfaceHolder.Callback, SeekBar.OnSee
 
     override fun onStartTrackingTouch(seekBar: SeekBar) {
         initMediaPlayer()
+        if (mMediaPlayer == null)
+            return
+
         mMediaPlayer!!.pause()
         mIsDragged = true
     }