+++ app/src/main/kotlin/com/simplemobiletools/gallery/helpers/Constants.kt
@@ -16,7 +16,13 @@ val DISPLAY_FILE_NAMES = "display_file_names"
 val DARK_BACKGROUND = "dark_background"
 val PINNED_FOLDERS = "pinned_folders"
 val DIR_COLUMN_CNT = "dir_column_cnt"
+val DIR_LANDSCAPE_COLUMN_CNT = "dir_landscape_column_cnt"
+val DIR_HORIZONTAL_COLUMN_CNT = "dir_horizontal_column_cnt"
+val DIR_LANDSCAPE_HORIZONTAL_COLUMN_CNT = "dir_landscape_horizontal_column_cnt"
 val MEDIA_COLUMN_CNT = "media_column_cnt"
+val MEDIA_LANDSCAPE_COLUMN_CNT = "media_landscape_column_cnt"
+val MEDIA_HORIZONTAL_COLUMN_CNT = "media_horizontal_column_cnt"
+val MEDIA_LANDSCAPE_HORIZONTAL_COLUMN_CNT = "media_landscape_horizontal_column_cnt"
 val SHOW_ALL = "show_all"                           // display images and videos from all folders together
 val SHOW_MEDIA = "show_media"
 val SAVE_FOLDER_PREFIX = "folder2_"
@@ -24,6 +30,7 @@ val HIDE_FOLDER_TOOLTIP_SHOWN = "hide_folder_tooltip_shown"
 val EXCLUDED_FOLDERS = "excluded_folders"
 val INCLUDED_FOLDERS = "included_folders"
 val ALBUM_COVERS = "album_covers"
+val SCROLL_HORIZONTALLY = "scroll_horizontally"
 
 val NOMEDIA = ".nomedia"
 
@@ -37,7 +44,7 @@ val DIRECTORIES = "directories2"
 val IS_VIEW_INTENT = "is_view_intent"
 
 val REQUEST_EDIT_IMAGE = 1
-val REQUEST_SET_WALLPAPER = 2
+val REQUEST_SET_AS = 2
 
 // show media
 val IMAGES_AND_VIDEOS = 0