+++ app/src/main/kotlin/com/simplemobiletools/gallery/views/MySquareImageView.kt
@@ -5,6 +5,7 @@ import android.util.AttributeSet
 import android.widget.ImageView
 
 class MySquareImageView : ImageView {
+    var isVerticalScrolling = true
 
     constructor(context: Context) : super(context)
 
@@ -14,6 +15,7 @@ class MySquareImageView : ImageView {
 
     override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
         super.onMeasure(widthMeasureSpec, heightMeasureSpec)
-        setMeasuredDimension(measuredWidth, measuredWidth)
+        val spec = if (isVerticalScrolling) measuredWidth else measuredHeight
+        setMeasuredDimension(spec, spec)
     }
 }